pyproject.toml
[project]
name = "pruebafapi"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "fastapi[standard]>=0.135.1",
    "numpy>=2.4.2",
    "pillow>=12.1.1",
    "pydantic-settings>=2.13.1",
    "python-multipart>=0.0.22",
    "requests>=2.32.5",
    "streamlit>=1.55.0",
    "tqdm>=4.67.3",
    "ultralytics>=8.4.21",
    "uvicorn>=0.41.0",
]
